+++ teamA/jeongHaeMan1/controllers/postController.js
@@ -1,5 +1,3 @@
-//controller/postController.js
-
 const postService = require('../services/postService');
 
 const write = async (req, res) => {
@@ -19,4 +17,48 @@ const write = async (req, res) => {
     }
 };
 
-module.exports = {write}
+const search = async (req, res, next) => {
+    try {
+        const search = await postService.search();
+
+        return res.status(201).json({ data : search });
+    } catch (err) {
+        console.log(err);
+        return res.status(err.statusCode || 500).json({ message : err.meassgae});
+    }
+};
+
+const modify = async (req, res) => {
+    try {
+        const {content, postId} = req.body;
+
+        if (  !content || !postId ) {
+            return res.status(400).json({ message : 'KEY_ERROR' });
+        }
+
+        const result = await postService.modify( content, postId );
+
+        res.status(201).json({ data : result });
+    }
+    catch (err) {
+        console.log(err);
+        return res.status(err.statusCode || 500).json({ message : err.meassgae});
+    }
+}
+
+const deletePost = async(req, res) => {
+    try{
+        const deleteId = Number(req.params.id);
+        await postService.deletePost(deleteId);
+
+        return res.status(200).json({data : 'postingDeleted'})
+    }
+    catch (err) {
+        console.error(err)
+        return res.status(err.statusCode || 500).json({message:  err.message})
+    }
+}
+
+module.exports = {
+    write, search, modify, deletePost
+}
