+++ teamA/jeongHaeMan1/models/likesDao.js
@@ -1,23 +1,5 @@
-//models/userDao.js
-const { DataSource } = require('typeorm');
+const myDataSource = require("../utils/typeorm");
 
-const myDataSource = new DataSource({
-    type: process.env.TYPEORM_CONNECTION,
-    username: process.env.TYPEORM_USERNAME,
-    password: process.env.TYPEORM_PASSWORD,
-    host: process.env.TYPEORM_HOST,
-    port: process.env.TYPEORM_PORT,
-    database: process.env.TYPEORM_DATABASE
-})
-
-myDataSource.initialize()
-    .then(() => {
-        console.log("Data Source has been initialized!");
-    })
-    .catch((err) => {
-        console.error("Error during Data Source initialization", err)
-        myDataSource.destroy()
-    })
 
 const likes = async (user_id, post_id) => {
     try {
