+++ teamA/jeongHaeMan1/models/postDao.js
@@ -1,23 +1,4 @@
-//models/userDao.js
-const { DataSource } = require('typeorm');
-
-const myDataSource = new DataSource({
-    type: process.env.TYPEORM_CONNECTION,
-    username: process.env.TYPEORM_USERNAME,
-    password: process.env.TYPEORM_PASSWORD,
-    host: process.env.TYPEORM_HOST,
-    port: process.env.TYPEORM_PORT,
-    database: process.env.TYPEORM_DATABASE
-})
-
-myDataSource.initialize()
-    .then(() => {
-        console.log("Data Source has been initialized!");
-    })
-    .catch((err) => {
-        console.error("Error during Data Source initialization", err)
-        myDataSource.destroy()
-    })
+const myDataSource = require("../utils/typeorm");
 
 const createPost = async (title, content, user_id, image_url) => {
     try {
@@ -38,6 +19,67 @@ const createPost = async (title, content, user_id, image_url) => {
     }
 };
 
+const createGetPost = async () => {
+    try {
+        return await myDataSource.query(
+        `SELECT 
+            users.id as userId, 
+            users.profile_image as userProfileImage, 
+            posts.id as postingId,
+            posts.image_url as postingImageUrl,
+            posts.content as postingContent
+        FROM (users, posts)
+        WHERE users.id = posts.user_id;
+        `)
+    } catch (err) {
+        const error = new Error(err.message);
+        error.statusCode = 500;
+        throw error;
+    }
+};
+
+const modifyPost = async (content, postId) => {
+    try {
+        await myDataSource.query(
+        `UPDATE posts SET
+            content = ?
+            WHERE id = ?`,
+        [content, postId]
+        )
+
+        return await myDataSource.query(
+            `SELECT 
+                users.id as userId, 
+                users.name as userName, 
+                posts.id as postingId,
+                posts.title as postingTitle,
+                posts.content as postingContent    
+            FROM (users, posts) 
+            WHERE users.id=posts.user_id AND posts.id= ${[postId]}`
+
+        )
+
+    } catch (err) {
+        const error = new Error(err.message);
+        error.statusCode = 500;
+        throw error;
+    }
+};
+
+const deletePost = async (deleteId) => {
+    try {
+        return await myDataSource.query(
+        `DELETE 
+            FROM posts
+        WHERE id = ${deleteId};
+        `)
+    } catch (err) {
+        const error = new Error(err.message);
+        error.statusCode = 500;
+        throw error;
+    }
+};
+
 module.exports = {
-    createPost
+    createPost, createGetPost, modifyPost, deletePost
 }
