--- teamA/jeongHaeMan1/routes/deletePostRouter.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const deletePostController = require('../controllers/deletePostController');
-
-const router = express.Router();
-
-router.delete("/delete/:id", deletePostController.deletePost);
-
-module.exports = {
-    router
-}
