--- teamA/jeongHaeMan1/routes/getPostRouter.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const getPostController = require('../controllers/getPostController');
-
-const router = express.Router();
-
-router.get('/search', getPostController.search);
-
-module.exports = {
-    router
-}
