--- teamA/jeongHaeMan1/routes/modifyPostRouter.js
@@ -1,10 +1,0 @@
-const express =  require('express');
-const modifyPostController = require("../controllers/modifyPostController");
-
-const router = express.Router();
-
-router.patch("/modify", modifyPostController.modify);
-
-module.exports = {
-    router
-}
