--- teamA/jeongHaeMan1/routes/selectUserRouter.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const selectUserController = require('../controllers/selectUserController');
-
-router.get('/getuser/:id', selectUserController.select);
-
-module.exports = {
-    router
-}
