+++ teamA/jeongHaeMan1/routes/userRouter.js
@@ -1,9 +1,10 @@
 const express = require('express');
-const userController = require('../controllers/userController');
-
 const router = express.Router();
 
+const userController = require('../controllers/userController');
+
 router.post('/signup', userController.signUp);
+router.get('/getuser/:id', userController.select);
 
 module.exports = {
     router
