--- teamA/jeongHaeMan1/services/getPostService.js
@@ -1,11 +1,0 @@
-const getPostDao = require('../models/getPostDao')
-
-const search = async () => {
-    const createGetPost = await getPostDao.createGetPost();
-
-    return createGetPost;
-};
-
-module.exports = {
-    search
-}
