--- teamA/jeongHaeMan1/services/selectUserService.js
@@ -1,12 +1,0 @@
-const selectUserDao = require('../models/selectUserDao');
-
-const select = async (userId) => {
-    const createSelectUser = await selectUserDao.createSelectUser(userId);
-
-    return createSelectUser;
-
-}
-
-module.exports = {
-    select
-}
