+++ teamA/jeongHaeMan1/utils/typeorm.js
@@ -1,0 +1,12 @@
+const { DataSource } = require('typeorm');
+
+const myDataSource = new DataSource({
+    type: process.env.TYPEORM_CONNECTION,
+    username: process.env.TYPEORM_USERNAME,
+    password: process.env.TYPEORM_PASSWORD,
+    host: process.env.TYPEORM_HOST,
+    port: process.env.TYPEORM_PORT,
+    database: process.env.TYPEORM_DATABASE
+})
+
+module.exports = myDataSource;
